MPSC for locks, fixing MPSC node in processors.
The locks are now implemented with the MPSC queue, the nodes that
were intrusive in the processors won't work now because the tasks
already were intrusive. So now we just put a regular node in there,
and manually keep the contents in sync.

The benefits are still good: the queues that hold the processors
in theory shouldn't need any extra allocations.