Adding a few more exprs, changing class type rep.
The class type can't really be a structural thing, as it may be recursive and
we have to perform structural equality on it to get `test :: Eq a a` to work.
Therefore we fall back to just the name with an open-ended `Typ` representation.

It isn't beautiful but I can't think of a more straight-forward way to do it.